move lehmer into utils
